Remove InsecureTrustManagerFactory
